// apps/web/src/generated/requests/types.gen.ts
export type UpdateVenueDto = {
  name?: string;
  typeId?: string;
  floor?: string;
  area?: number;
  pricePerHour?: number;
  isActive?: boolean;
};